+++ src/components/ImageTile.tsx
@@ -1,33 +1,10 @@
 import Image from 'next/image';
-import Link from 'next/link';
-import { useContext } from 'react';
 
 import { BlogImage } from 'src/app/interfaces_blog';
+import { ImageTileClientWrapper } from 'src/components/ImageTileClientWrapper';
 import { urlFor } from 'src/lib/imageHelpers';
-import {
-  CurrentImageIndexContext,
-  ImageListContext,
-  LoadImageContext
-} from 'src/providers/ImageContext';
+import { LoadImageContext } from 'src/providers/ImageContext';
 
-const ImageTileClientWrapper = ({ children, index }: { children: React.ReactNode; index: number }) => {
-  const [currentImageIdList, ] = useContext(ImageListContext);
-  const [, setCurrentImageIndex] = useContext(CurrentImageIndexContext);
-
-  const setImgIndex = () => {
-    setCurrentImageIndex(index);
-  };
-
-  return (
-    <Link href={ `/photos/${currentImageIdList[index]}` }
-      onClick={ setImgIndex }
-      className='flex grow w-60 h-auto rounded ease-out transition-all duration-[250ms]
-      hover:scale-[1.02] hover:shadow-[0_4px_8px_rgba(0,0,0,0.3)]'
-    >
-      { children }
-    </Link>
-  );
-};
 const ImageTile = ({ photo, index }: { photo: BlogImage; index: number }) => (
   <ImageTileClientWrapper index={ index }>
     <Image
